+++ packages/test-utils/src/lib.rs
@@ -1,3 +1,6 @@
+#[cfg(not(target_arch = "wasm32"))]
+pub mod migration;
+
 use babylon_apis::btc_staking_api::{
     ActiveBtcDelegation, BtcUndelegationInfo, CovenantAdaptorSignatures, DelegatorUnbondingInfo,
     NewFinalityProvider, ProofOfPossessionBtc,